requirements.txt
aws_cdk.core==1.138.1
aws_cdk.aws_iam==1.138.1
virtualenv
